+++ build.zig
@@ -42,6 +42,9 @@ pub fn build(b: *std.Build) void {
 
     const run_step = b.step("run", "Run the app");
     const run_cmd = b.addRunArtifact(repl);
+    if (b.args) |args| {
+        run_cmd.addArgs(args);
+    }
     run_step.dependOn(&run_cmd.step);
     run_cmd.step.dependOn(b.getInstallStep());